pyproject.toml
[tool.poetry]
name = "dsn"
version = "0.1.0"
description = ""
authors = ["RenChu Wang <renchu@github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.1"
scikit-learn = "^1.3.2"
vadersentiment = "^3.3.2"
pillow = "^10.1.0"
numpy = "^1.26.2"
scipy = "^1.11.4"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
